keep filter state in browser query string
by having a local state with the keys. on state changes the browser query string is and hisory is updated using pushState.
When the user navigates back and forth in the hisory the popstate event is fired and the new browser query string is written to the local state.

Had the popstate event been fired when the browser history is changed using the pushState function, it would not be neseccary to have the local state at all. Unfortunately that is not the case.